// src/api/generated/models/GetOptionResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GetModelResultDto } from './GetModelResultDto';
export type GetOptionResultDto = {
  _id: string;
  optionName: string;
  active: boolean;
  createBy: string;
  createAt: string;
  modelId: GetModelResultDto;
};