// src/api/generated/models/ProductResultGetDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
import type { GetOptionResultDto } from './GetOptionResultDto';
export type ProductResultGetDto = {
  _id: string;
  productName: string;
  size: string;
  group: string;
  cost: number;
  sellingPrice: number;
  active: boolean;
  optionId: GetOptionResultDto;
};